+++ test/grid/__init__.py
@@ -1,0 +1,1 @@
+"""Integration test module."""
